feat(lib/lv_bindings): add m5stack support
This commit updates the `lib/lv_bindings` submodule to add optional support for
[M5Stack cores][m5cores] using the ILI9341 LCD display (e.g., [this one][m5]).
Without this commit, the display is not properly configured resulting in at
least the following issues:

 - horizontally mirrored image
 - inverted color

To enable M5Stack support, define `M5STACK_ILI9341`, e.g.:

    make -C ports/esp32 LV_CFLAGS="-DM5STACK_ILI9341 -DLV_COLOR_DEPTH=16" BOARD=GENERIC_SPIRAM PYTHON=python2 deploy

Also, if `M5STACK_ILI9341` is defined, include the Roboto 12pt font and select
it as the default font to accommodate the small ILI9341 320x240 LCD resolution.

See [sci-bots/m5-lvgl][m5-lvgl] for a MicroPython driver for the M5Stack;
including an input driver using the M5Stack A, B, C buttons to emulate an
encoder input device. Here is a video demonstrating the functionality:

[![][demo-thumbnail]][demo-video]

[demo-video]: https://www.youtube.com/watch?v=lpJi6oqPu_4&feature=youtu.be
[demo-thumbnail]: https://trello-attachments.s3.amazonaws.com/5a4e5802bd8b7c317195569d/5dd6d56a74cecc832ce7b61f/5798274fae2f7e186dc2761263dbcd7f/image.png
[m5cores]: https://m5stack.com/collections/m5-core
[m5]: https://m5stack.com/collections/m5-core/products/basic-core-iot-development-kit
[m5-lvgl]: https://github.com/sci-bots/m5-lvgl